Add files via upload
Part 1 and two are done, part 4 works but im not sure the forum has the right stuff on it, im pretty sure it does but we can change it easily if needed. I did not start part 5. As for part 4 i got the adapter made and the it pulls data from the firebase. Look at older commits for info on other assignments for the week.